refactor: separar modelos en sus submodulos
