+++ backend/models/estados.py
@@ -1,0 +1,5 @@
+from mongoengine import Document, StringField
+
+
+class Estados(Document):
+    nombre = StringField(unique=True)
