+++ backend/models/roles.py
@@ -1,0 +1,5 @@
+from mongoengine import Document, StringField
+
+
+class Roles(Document):
+    nombre = StringField(unique=True)
